requirements.txt
# Palantiny Web Server
# FastAPI 및 ASGI 서버
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 데이터베이스 (PostgreSQL 비동기)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# 설정 및 환경변수
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# JWT (MVP 로그인)
python-jose[cryptography]>=3.3.0

# Neo4j AuraDB
neo4j>=5.0.0

# 비밀번호 해시
passlib[bcrypt]>=1.7.4

# DJMEDI 외부 API 호출
httpx>=0.27.0
